Standardize on full names with underscore separators
Before, there was a mix of short names like "dic", and long names like
"iterables". There was a mix of underscores like "type_changes" and no
underscores like "newtype". I tried to find all of these cases and come
up with a consistent usage of long, descriptive names that are separated
by underscores.