Clean up docstrings to match PEP257 recommendation
From PEP257: For consistency, always use """triple double quotes"""
around docstrings.

There were also a couple of other errors that flake8 pointed out along
with a quadruple quote on one docstring that python didn't seem to mind,
but was still wrong. And one docstring attempt only had single quotes
which means it wasn't really a docstring, but just a string that was
declared and never used.