This allows for a custom compare function to compare iterable items.
The new compare function takes two items of an iterable and should return True
it matching, False if no match, and raise CannotCompare if unable to compare the two items.

The default behavior is the same as before which is comparing each item in order.

If the compare function raises CannotCompare then behavior reverts back to the default in order.

This also introduces a new report key which is `iterable_item_moved` to track if iterable items have moved.