Fix for diffing using iterable_compare_func with nested objects.
This commit addresses two issues. First ensuring that the diff indexes
for moved items are always relative to t2 (except for removed)
to stay consistent with the rest of the diff types.

Second, when replaying moved items ensure that the new values is
replaced after adding the items. Since the moved items already have any
nested items inside of them, there is no need to replay those nested
added items (it was causing items to get double added).