Move from legacy setup.py to pyproject.toml with flit
This change simplifies the build logic on the new packaging metadata
format provided with `pyproject.toml` using the flit build backend. The
setuptools build backend wasn't featureful enough to be usable.

This still doesn't fix the fact that installing `deepdiff` results in a
broken `deep` CLI command, but it at least pushes the ball towards a
world where that will be possible, someday.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>